+++ erpnext/controllers/queries.py
@@ -293,3 +293,24 @@ def get_account_list(doctype, txt, searchfield, start, page_len, filters):
 		fields = ["name", "parent_account"],
 		limit_start=start, limit_page_length=page_len, as_list=True)
 
+
+@frappe.whitelist()
+def get_income_account(doctype, txt, searchfield, start, page_len, filters):
+	from erpnext.controllers.queries import get_match_cond
+
+	# income account can be any Credit account,
+	# but can also be a Asset account with account_type='Income Account' in special circumstances.
+	# Hence the first condition is an "OR"
+	condition = ""
+	if filters.get("company"):
+		condition += "and tabAccount.company = '%(company)s'"
+	
+	return frappe.db.sql("""select tabAccount.name from `tabAccount`
+			where (tabAccount.report_type = "Profit and Loss"
+					or tabAccount.account_type in ("Income Account", "Temporary"))
+				and tabAccount.is_group=0
+				and tabAccount.docstatus!=2
+				and tabAccount.%(key)s LIKE '%(txt)s'
+				{condition} {match_condition}"""
+			.format(condition=condition, match_condition=get_match_cond(doctype)),
+			{'key': searchfield, 'txt': "%%%s%%" % frappe.db.escape(txt)})
